Moved all Format implementations to their own file
