+++ go/vt/sqlparser/ast.go
@@ -271,7 +271,8 @@ func Append(buf *strings.Builder, node SQLNode) {
 	node.Format(tbuf)
 }
 
-type ( // Statements
+// Statements
+type (
 	// Statement represents a statement.
 	Statement interface {
 		iStatement()
@@ -502,15 +503,6 @@ func (node *Select) SetLimit(limit *Limit) {
 	node.Limit = limit
 }
 
-// Format formats the node.
-func (node *Select) Format(buf *TrackedBuffer) {
-	buf.Myprintf("select %v%s%s%s%v from %v%v%v%v%v%v%s",
-		node.Comments, node.Cache, node.Distinct, node.Hints, node.SelectExprs,
-		node.From, node.Where,
-		node.GroupBy, node.Having, node.OrderBy,
-		node.Limit, node.Lock)
-}
-
 func (node *Select) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -582,11 +574,6 @@ func (node *ParenSelect) SetLimit(limit *Limit) {
 	panic("unreachable")
 }
 
-// Format formats the node.
-func (node *ParenSelect) Format(buf *TrackedBuffer) {
-	buf.Myprintf("(%v)", node.Select)
-}
-
 func (node *ParenSelect) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -614,12 +601,6 @@ func (node *Union) SetLimit(limit *Limit) {
 	node.Limit = limit
 }
 
-// Format formats the node.
-func (node *Union) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%v %s %v%v%v%s", node.Left, node.Type, node.Right,
-		node.OrderBy, node.Limit, node.Lock)
-}
-
 func (node *Union) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -629,12 +610,6 @@ func (node *Union) walkSubtree(visit Visit) error {
 		node.Left,
 		node.Right,
 	)
-}
-
-// Format formats the node.
-func (node *Stream) Format(buf *TrackedBuffer) {
-	buf.Myprintf("stream %v%v from %v",
-		node.Comments, node.SelectExpr, node.Table)
 }
 
 func (node *Stream) walkSubtree(visit Visit) error {
@@ -654,14 +629,6 @@ const (
 	InsertStr  = "insert"
 	ReplaceStr = "replace"
 )
-
-// Format formats the node.
-func (node *Insert) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%s %v%sinto %v%v%v %v%v",
-		node.Action,
-		node.Comments, node.Ignore,
-		node.Table, node.Partitions, node.Columns, node.Rows, node.OnDup)
-}
 
 func (node *Insert) walkSubtree(visit Visit) error {
 	if node == nil {
@@ -688,13 +655,6 @@ func (*Union) iInsertRows()       {}
 func (Values) iInsertRows()       {}
 func (*ParenSelect) iInsertRows() {}
 
-// Format formats the node.
-func (node *Update) Format(buf *TrackedBuffer) {
-	buf.Myprintf("update %v%s%v set %v%v%v%v",
-		node.Comments, node.Ignore, node.TableExprs,
-		node.Exprs, node.Where, node.OrderBy, node.Limit)
-}
-
 func (node *Update) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -708,15 +668,6 @@ func (node *Update) walkSubtree(visit Visit) error {
 		node.OrderBy,
 		node.Limit,
 	)
-}
-
-// Format formats the node.
-func (node *Delete) Format(buf *TrackedBuffer) {
-	buf.Myprintf("delete %v", node.Comments)
-	if node.Targets != nil {
-		buf.Myprintf("%v ", node.Targets)
-	}
-	buf.Myprintf("from %v%v%v%v%v", node.TableExprs, node.Partitions, node.Where, node.OrderBy, node.Limit)
 }
 
 func (node *Delete) walkSubtree(visit Visit) error {
@@ -742,15 +693,6 @@ const (
 	ImplicitStr       = ""
 )
 
-// Format formats the node.
-func (node *Set) Format(buf *TrackedBuffer) {
-	if node.Scope == "" {
-		buf.Myprintf("set %v%v", node.Comments, node.Exprs)
-	} else {
-		buf.Myprintf("set %v%s %v", node.Comments, node.Scope, node.Exprs)
-	}
-}
-
 func (node *Set) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -760,20 +702,6 @@ func (node *Set) walkSubtree(visit Visit) error {
 		node.Comments,
 		node.Exprs,
 	)
-}
-
-// Format formats the node.
-func (node *DBDDL) Format(buf *TrackedBuffer) {
-	switch node.Action {
-	case CreateStr:
-		buf.WriteString(fmt.Sprintf("%s database %s", node.Action, node.DBName))
-	case DropStr:
-		exists := ""
-		if node.IfExists {
-			exists = " if exists"
-		}
-		buf.WriteString(fmt.Sprintf("%s database%s %v", node.Action, exists, node.DBName))
-	}
 }
 
 // walkSubtree walks the nodes of the subtree.
@@ -801,68 +729,6 @@ const (
 	// Vindex DDL param to specify the owner of a vindex
 	VindexOwnerStr = "owner"
 )
-
-// Format formats the node.
-func (node *DDL) Format(buf *TrackedBuffer) {
-	switch node.Action {
-	case CreateStr:
-		if node.OptLike != nil {
-			buf.Myprintf("%s table %v %v", node.Action, node.Table, node.OptLike)
-		} else if node.TableSpec != nil {
-			buf.Myprintf("%s table %v %v", node.Action, node.Table, node.TableSpec)
-		} else {
-			buf.Myprintf("%s table %v", node.Action, node.Table)
-		}
-	case DropStr:
-		exists := ""
-		if node.IfExists {
-			exists = " if exists"
-		}
-		buf.Myprintf("%s table%s %v", node.Action, exists, node.FromTables)
-	case RenameStr:
-		buf.Myprintf("%s table %v to %v", node.Action, node.FromTables[0], node.ToTables[0])
-		for i := 1; i < len(node.FromTables); i++ {
-			buf.Myprintf(", %v to %v", node.FromTables[i], node.ToTables[i])
-		}
-	case AlterStr:
-		if node.PartitionSpec != nil {
-			buf.Myprintf("%s table %v %v", node.Action, node.Table, node.PartitionSpec)
-		} else {
-			buf.Myprintf("%s table %v", node.Action, node.Table)
-		}
-	case FlushStr:
-		buf.Myprintf("%s", node.Action)
-	case CreateVindexStr:
-		buf.Myprintf("alter vschema create vindex %v %v", node.Table, node.VindexSpec)
-	case DropVindexStr:
-		buf.Myprintf("alter vschema drop vindex %v", node.Table)
-	case AddVschemaTableStr:
-		buf.Myprintf("alter vschema add table %v", node.Table)
-	case DropVschemaTableStr:
-		buf.Myprintf("alter vschema drop table %v", node.Table)
-	case AddColVindexStr:
-		buf.Myprintf("alter vschema on %v add vindex %v (", node.Table, node.VindexSpec.Name)
-		for i, col := range node.VindexCols {
-			if i != 0 {
-				buf.Myprintf(", %v", col)
-			} else {
-				buf.Myprintf("%v", col)
-			}
-		}
-		buf.Myprintf(")")
-		if node.VindexSpec.Type.String() != "" {
-			buf.Myprintf(" %v", node.VindexSpec)
-		}
-	case DropColVindexStr:
-		buf.Myprintf("alter vschema on %v drop vindex %v", node.Table, node.VindexSpec.Name)
-	case AddSequenceStr:
-		buf.Myprintf("alter vschema add sequence %v", node.Table)
-	case AddAutoIncStr:
-		buf.Myprintf("alter vschema on %v add auto_increment %v", node.Table, node.AutoIncSpec)
-	default:
-		buf.Myprintf("%s table %v", node.Action, node.Table)
-	}
-}
 
 func (node *DDL) walkSubtree(visit Visit) error {
 	if node == nil {
@@ -897,11 +763,6 @@ type OptLike struct {
 	LikeTable TableName
 }
 
-// Format formats the node.
-func (node *OptLike) Format(buf *TrackedBuffer) {
-	buf.Myprintf("like %v", node.LikeTable)
-}
-
 func (node *OptLike) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -914,22 +775,6 @@ type PartitionSpec struct {
 	Action      string
 	Name        ColIdent
 	Definitions []*PartitionDefinition
-}
-
-// Format formats the node.
-func (node *PartitionSpec) Format(buf *TrackedBuffer) {
-	switch node.Action {
-	case ReorganizeStr:
-		buf.Myprintf("%s %v into (", node.Action, node.Name)
-		var prefix string
-		for _, pd := range node.Definitions {
-			buf.Myprintf("%s%v", prefix, pd)
-			prefix = ", "
-		}
-		buf.Myprintf(")")
-	default:
-		panic("unimplemented")
-	}
 }
 
 func (node *PartitionSpec) walkSubtree(visit Visit) error {
@@ -954,15 +799,6 @@ type PartitionDefinition struct {
 	Maxvalue bool
 }
 
-// Format formats the node
-func (node *PartitionDefinition) Format(buf *TrackedBuffer) {
-	if !node.Maxvalue {
-		buf.Myprintf("partition %v values less than (%v)", node.Name, node.Limit)
-	} else {
-		buf.Myprintf("partition %v values less than (maxvalue)", node.Name)
-	}
-}
-
 func (node *PartitionDefinition) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -980,26 +816,6 @@ type TableSpec struct {
 	Indexes     []*IndexDefinition
 	Constraints []*ConstraintDefinition
 	Options     string
-}
-
-// Format formats the node.
-func (ts *TableSpec) Format(buf *TrackedBuffer) {
-	buf.Myprintf("(\n")
-	for i, col := range ts.Columns {
-		if i == 0 {
-			buf.Myprintf("\t%v", col)
-		} else {
-			buf.Myprintf(",\n\t%v", col)
-		}
-	}
-	for _, idx := range ts.Indexes {
-		buf.Myprintf(",\n\t%v", idx)
-	}
-	for _, c := range ts.Constraints {
-		buf.Myprintf(",\n\t%v", c)
-	}
-
-	buf.Myprintf("\n)%s", strings.Replace(ts.Options, ", ", ",\n  ", -1))
 }
 
 // AddColumn appends the given column to the list in the spec
@@ -1049,11 +865,6 @@ type ColumnDefinition struct {
 	Type ColumnType
 }
 
-// Format formats the node.
-func (col *ColumnDefinition) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%v %v", col.Name, &col.Type)
-}
-
 func (col *ColumnDefinition) walkSubtree(visit Visit) error {
 	if col == nil {
 		return nil
@@ -1093,70 +904,6 @@ type ColumnType struct {
 
 	// Key specification
 	KeyOpt ColumnKeyOption
-}
-
-// Format returns a canonical string representation of the type and all relevant options
-func (ct *ColumnType) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%s", ct.Type)
-
-	if ct.Length != nil && ct.Scale != nil {
-		buf.Myprintf("(%v,%v)", ct.Length, ct.Scale)
-
-	} else if ct.Length != nil {
-		buf.Myprintf("(%v)", ct.Length)
-	}
-
-	if ct.EnumValues != nil {
-		buf.Myprintf("(%s)", strings.Join(ct.EnumValues, ", "))
-	}
-
-	opts := make([]string, 0, 16)
-	if ct.Unsigned {
-		opts = append(opts, keywordStrings[UNSIGNED])
-	}
-	if ct.Zerofill {
-		opts = append(opts, keywordStrings[ZEROFILL])
-	}
-	if ct.Charset != "" {
-		opts = append(opts, keywordStrings[CHARACTER], keywordStrings[SET], ct.Charset)
-	}
-	if ct.Collate != "" {
-		opts = append(opts, keywordStrings[COLLATE], ct.Collate)
-	}
-	if ct.NotNull {
-		opts = append(opts, keywordStrings[NOT], keywordStrings[NULL])
-	}
-	if ct.Default != nil {
-		opts = append(opts, keywordStrings[DEFAULT], String(ct.Default))
-	}
-	if ct.OnUpdate != nil {
-		opts = append(opts, keywordStrings[ON], keywordStrings[UPDATE], String(ct.OnUpdate))
-	}
-	if ct.Autoincrement {
-		opts = append(opts, keywordStrings[AUTO_INCREMENT])
-	}
-	if ct.Comment != nil {
-		opts = append(opts, keywordStrings[COMMENT_KEYWORD], String(ct.Comment))
-	}
-	if ct.KeyOpt == colKeyPrimary {
-		opts = append(opts, keywordStrings[PRIMARY], keywordStrings[KEY])
-	}
-	if ct.KeyOpt == colKeyUnique {
-		opts = append(opts, keywordStrings[UNIQUE])
-	}
-	if ct.KeyOpt == colKeyUniqueKey {
-		opts = append(opts, keywordStrings[UNIQUE], keywordStrings[KEY])
-	}
-	if ct.KeyOpt == colKeySpatialKey {
-		opts = append(opts, keywordStrings[SPATIAL], keywordStrings[KEY])
-	}
-	if ct.KeyOpt == colKey {
-		opts = append(opts, keywordStrings[KEY])
-	}
-
-	if len(opts) != 0 {
-		buf.Myprintf(" %s", strings.Join(opts, " "))
-	}
 }
 
 // DescribeType returns the abbreviated type information as required for
@@ -1294,31 +1041,6 @@ type IndexDefinition struct {
 	Options []*IndexOption
 }
 
-// Format formats the node.
-func (idx *IndexDefinition) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%v (", idx.Info)
-	for i, col := range idx.Columns {
-		if i != 0 {
-			buf.Myprintf(", %v", col.Column)
-		} else {
-			buf.Myprintf("%v", col.Column)
-		}
-		if col.Length != nil {
-			buf.Myprintf("(%v)", col.Length)
-		}
-	}
-	buf.Myprintf(")")
-
-	for _, opt := range idx.Options {
-		buf.Myprintf(" %s", opt.Name)
-		if opt.Using != "" {
-			buf.Myprintf(" %s", opt.Using)
-		} else {
-			buf.Myprintf(" %v", opt.Value)
-		}
-	}
-}
-
 func (idx *IndexDefinition) walkSubtree(visit Visit) error {
 	if idx == nil {
 		return nil
@@ -1340,18 +1062,6 @@ type IndexInfo struct {
 	Primary bool
 	Spatial bool
 	Unique  bool
-}
-
-// Format formats the node.
-func (ii *IndexInfo) Format(buf *TrackedBuffer) {
-	if ii.Primary {
-		buf.Myprintf("%s", ii.Type)
-	} else {
-		buf.Myprintf("%s", ii.Type)
-		if !ii.Name.IsEmpty() {
-			buf.Myprintf(" %v", ii.Name)
-		}
-	}
 }
 
 func (ii *IndexInfo) walkSubtree(visit Visit) error {
@@ -1404,12 +1114,6 @@ type AutoIncSpec struct {
 	Sequence TableName
 }
 
-// Format formats the node.
-func (node *AutoIncSpec) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%v ", node.Column)
-	buf.Myprintf("using %v", node.Sequence)
-}
-
 func (node *AutoIncSpec) walkSubtree(visit Visit) error {
 	err := Walk(visit, node.Sequence, node.Column)
 	return err
@@ -1428,24 +1132,6 @@ func (node *VindexSpec) ParseParams() (string, map[string]string) {
 		}
 	}
 	return owner, params
-}
-
-// Format formats the node. The "CREATE VINDEX" preamble was formatted in
-// the containing DDL node Format, so this just prints the type, any
-// parameters, and optionally the owner
-func (node *VindexSpec) Format(buf *TrackedBuffer) {
-	buf.Myprintf("using %v", node.Type)
-
-	numParams := len(node.Params)
-	if numParams != 0 {
-		buf.Myprintf(" with ")
-		for i, p := range node.Params {
-			if i != 0 {
-				buf.Myprintf(", ")
-			}
-			buf.Myprintf("%v", p)
-		}
-	}
 }
 
 func (node *VindexSpec) walkSubtree(visit Visit) error {
@@ -1473,11 +1159,6 @@ type VindexParam struct {
 	Val string
 }
 
-// Format formats the node.
-func (node VindexParam) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%s=%s", node.Key.String(), node.Val)
-}
-
 func (node VindexParam) walkSubtree(visit Visit) error {
 	return Walk(visit,
 		node.Key,
@@ -1494,14 +1175,6 @@ type ConstraintDefinition struct {
 type ConstraintInfo interface {
 	SQLNode
 	constraintInfo()
-}
-
-// Format formats the node.
-func (c *ConstraintDefinition) Format(buf *TrackedBuffer) {
-	if c.Name != "" {
-		buf.Myprintf("constraint %s ", c.Name)
-	}
-	c.Details.Format(buf)
 }
 
 func (c *ConstraintDefinition) walkSubtree(visit Visit) error {
@@ -1526,22 +1199,6 @@ const (
 
 func (a ReferenceAction) walkSubtree(visit Visit) error { return nil }
 
-// Format formats the node.
-func (a ReferenceAction) Format(buf *TrackedBuffer) {
-	switch a {
-	case Restrict:
-		buf.WriteString("restrict")
-	case Cascade:
-		buf.WriteString("cascade")
-	case NoAction:
-		buf.WriteString("no action")
-	case SetNull:
-		buf.WriteString("set null")
-	case SetDefault:
-		buf.WriteString("set default")
-	}
-}
-
 // ForeignKeyDefinition describes a foreign key in a CREATE TABLE statement
 type ForeignKeyDefinition struct {
 	Source            Columns
@@ -1553,17 +1210,6 @@ type ForeignKeyDefinition struct {
 
 var _ ConstraintInfo = &ForeignKeyDefinition{}
 
-// Format formats the node.
-func (f *ForeignKeyDefinition) Format(buf *TrackedBuffer) {
-	buf.Myprintf("foreign key %v references %v %v", f.Source, f.ReferencedTable, f.ReferencedColumns)
-	if f.OnDelete != DefaultAction {
-		buf.Myprintf(" on delete %v", f.OnDelete)
-	}
-	if f.OnUpdate != DefaultAction {
-		buf.Myprintf(" on update %v", f.OnUpdate)
-	}
-}
-
 func (f *ForeignKeyDefinition) constraintInfo() {}
 
 func (f *ForeignKeyDefinition) walkSubtree(visit Visit) error {
@@ -1574,36 +1220,6 @@ func (f *ForeignKeyDefinition) walkSubtree(visit Visit) error {
 		return err
 	}
 	return Walk(visit, f.ReferencedColumns)
-}
-
-// Format formats the node.
-func (node *Show) Format(buf *TrackedBuffer) {
-	if (node.Type == "tables" || node.Type == "columns" || node.Type == "fields") && node.ShowTablesOpt != nil {
-		opt := node.ShowTablesOpt
-		buf.Myprintf("show %s%s", opt.Full, node.Type)
-		if (node.Type == "columns" || node.Type == "fields") && node.HasOnTable() {
-			buf.Myprintf(" from %v", node.OnTable)
-		}
-		if opt.DbName != "" {
-			buf.Myprintf(" from %s", opt.DbName)
-		}
-		buf.Myprintf("%v", opt.Filter)
-		return
-	}
-	if node.Scope == "" {
-		buf.Myprintf("show %s", node.Type)
-	} else {
-		buf.Myprintf("show %s %s", node.Scope, node.Type)
-	}
-	if node.HasOnTable() {
-		buf.Myprintf(" on %v", node.OnTable)
-	}
-	if node.Type == "collation" && node.ShowCollationFilterOpt != nil {
-		buf.Myprintf(" where %v", *node.ShowCollationFilterOpt)
-	}
-	if node.HasTable() {
-		buf.Myprintf(" %v", node.Table)
-	}
 }
 
 // HasOnTable returns true if the show statement has an "on" clause
@@ -1634,74 +1250,28 @@ type ShowFilter struct {
 	Filter Expr
 }
 
-// Format formats the node.
-func (node *ShowFilter) Format(buf *TrackedBuffer) {
-	if node == nil {
-		return
-	}
-	if node.Like != "" {
-		buf.Myprintf(" like '%s'", node.Like)
-	} else {
-		buf.Myprintf(" where %v", node.Filter)
-	}
-}
-
 func (node *ShowFilter) walkSubtree(visit Visit) error {
 	return nil
-}
-
-// Format formats the node.
-func (node *Use) Format(buf *TrackedBuffer) {
-	if node.DBName.v != "" {
-		buf.Myprintf("use %v", node.DBName)
-	} else {
-		buf.Myprintf("use")
-	}
 }
 
 func (node *Use) walkSubtree(visit Visit) error {
 	return Walk(visit, node.DBName)
 }
 
-// Format formats the node.
-func (node *Begin) Format(buf *TrackedBuffer) {
-	buf.WriteString("begin")
-}
-
 func (node *Begin) walkSubtree(visit Visit) error {
 	return nil
-}
-
-// Format formats the node.
-func (node *Commit) Format(buf *TrackedBuffer) {
-	buf.WriteString("commit")
 }
 
 func (node *Commit) walkSubtree(visit Visit) error {
 	return nil
 }
 
-// Format formats the node.
-func (node *Rollback) Format(buf *TrackedBuffer) {
-	buf.WriteString("rollback")
-}
-
 func (node *Rollback) walkSubtree(visit Visit) error {
 	return nil
 }
 
-// Format formats the node.
-func (node *OtherRead) Format(buf *TrackedBuffer) {
-	buf.WriteString("otherread")
-}
-
 func (node *OtherRead) walkSubtree(visit Visit) error {
 	return nil
-}
-
-// Format formats the node.
-func (node *OtherAdmin) Format(buf *TrackedBuffer) {
-	buf.WriteString("otheradmin")
 }
 
 func (node *OtherAdmin) walkSubtree(visit Visit) error {
@@ -1711,28 +1281,12 @@ func (node *OtherAdmin) walkSubtree(visit Visit) error {
 // Comments represents a list of comments.
 type Comments [][]byte
 
-// Format formats the node.
-func (node Comments) Format(buf *TrackedBuffer) {
-	for _, c := range node {
-		buf.Myprintf("%s ", c)
-	}
-}
-
 func (node Comments) walkSubtree(visit Visit) error {
 	return nil
 }
 
 // SelectExprs represents SELECT expressions.
 type SelectExprs []SelectExpr
-
-// Format formats the node.
-func (node SelectExprs) Format(buf *TrackedBuffer) {
-	var prefix string
-	for _, n := range node {
-		buf.Myprintf("%s%v", prefix, n)
-		prefix = ", "
-	}
-}
 
 func (node SelectExprs) walkSubtree(visit Visit) error {
 	for _, n := range node {
@@ -1758,14 +1312,6 @@ type StarExpr struct {
 	TableName TableName
 }
 
-// Format formats the node.
-func (node *StarExpr) Format(buf *TrackedBuffer) {
-	if !node.TableName.IsEmpty() {
-		buf.Myprintf("%v.", node.TableName)
-	}
-	buf.Myprintf("*")
-}
-
 func (node *StarExpr) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -1780,14 +1326,6 @@ func (node *StarExpr) walkSubtree(visit Visit) error {
 type AliasedExpr struct {
 	Expr Expr
 	As   ColIdent
-}
-
-// Format formats the node.
-func (node *AliasedExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%v", node.Expr)
-	if !node.As.IsEmpty() {
-		buf.Myprintf(" as %v", node.As)
-	}
 }
 
 func (node *AliasedExpr) walkSubtree(visit Visit) error {
@@ -1806,30 +1344,12 @@ type Nextval struct {
 	Expr Expr
 }
 
-// Format formats the node.
-func (node Nextval) Format(buf *TrackedBuffer) {
-	buf.Myprintf("next %v values", node.Expr)
-}
-
 func (node Nextval) walkSubtree(visit Visit) error {
 	return Walk(visit, node.Expr)
 }
 
 // Columns represents an insert column list.
 type Columns []ColIdent
-
-// Format formats the node.
-func (node Columns) Format(buf *TrackedBuffer) {
-	if node == nil {
-		return
-	}
-	prefix := "("
-	for _, n := range node {
-		buf.Myprintf("%s%v", prefix, n)
-		prefix = ", "
-	}
-	buf.WriteString(")")
-}
 
 func (node Columns) walkSubtree(visit Visit) error {
 	for _, n := range node {
@@ -1854,19 +1374,6 @@ func (node Columns) FindColumn(col ColIdent) int {
 // Partitions is a type alias for Columns so we can handle printing efficiently
 type Partitions Columns
 
-// Format formats the node
-func (node Partitions) Format(buf *TrackedBuffer) {
-	if node == nil {
-		return
-	}
-	prefix := " partition ("
-	for _, n := range node {
-		buf.Myprintf("%s%v", prefix, n)
-		prefix = ", "
-	}
-	buf.WriteString(")")
-}
-
 func (node Partitions) walkSubtree(visit Visit) error {
 	for _, n := range node {
 		if err := Walk(visit, n); err != nil {
@@ -1878,15 +1385,6 @@ func (node Partitions) walkSubtree(visit Visit) error {
 
 // TableExprs represents a list of table expressions.
 type TableExprs []TableExpr
-
-// Format formats the node.
-func (node TableExprs) Format(buf *TrackedBuffer) {
-	var prefix string
-	for _, n := range node {
-		buf.Myprintf("%s%v", prefix, n)
-		prefix = ", "
-	}
-}
 
 func (node TableExprs) walkSubtree(visit Visit) error {
 	for _, n := range node {
@@ -1915,18 +1413,6 @@ type AliasedTableExpr struct {
 	Partitions Partitions
 	As         TableIdent
 	Hints      *IndexHints
-}
-
-// Format formats the node.
-func (node *AliasedTableExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%v%v", node.Expr, node.Partitions)
-	if !node.As.IsEmpty() {
-		buf.Myprintf(" as %v", node.As)
-	}
-	if node.Hints != nil {
-		// Hint node provides the space padding.
-		buf.Myprintf("%v", node.Hints)
-	}
 }
 
 func (node *AliasedTableExpr) walkSubtree(visit Visit) error {
@@ -1960,15 +1446,6 @@ func (*Subquery) iSimpleTableExpr() {}
 // TableNames is a list of TableName.
 type TableNames []TableName
 
-// Format formats the node.
-func (node TableNames) Format(buf *TrackedBuffer) {
-	var prefix string
-	for _, n := range node {
-		buf.Myprintf("%s%v", prefix, n)
-		prefix = ", "
-	}
-}
-
 func (node TableNames) walkSubtree(visit Visit) error {
 	for _, n := range node {
 		if err := Walk(visit, n); err != nil {
@@ -1985,17 +1462,6 @@ func (node TableNames) walkSubtree(visit Visit) error {
 // and a TableName can also be used as key in a map.
 type TableName struct {
 	Name, Qualifier TableIdent
-}
-
-// Format formats the node.
-func (node TableName) Format(buf *TrackedBuffer) {
-	if node.IsEmpty() {
-		return
-	}
-	if !node.Qualifier.IsEmpty() {
-		buf.Myprintf("%v.", node.Qualifier)
-	}
-	buf.Myprintf("%v", node.Name)
 }
 
 func (node TableName) walkSubtree(visit Visit) error {
@@ -2027,11 +1493,6 @@ type ParenTableExpr struct {
 	Exprs TableExprs
 }
 
-// Format formats the node.
-func (node *ParenTableExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("(%v)", node.Exprs)
-}
-
 func (node *ParenTableExpr) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -2047,16 +1508,6 @@ func (node *ParenTableExpr) walkSubtree(visit Visit) error {
 type JoinCondition struct {
 	On    Expr
 	Using Columns
-}
-
-// Format formats the node.
-func (node JoinCondition) Format(buf *TrackedBuffer) {
-	if node.On != nil {
-		buf.Myprintf(" on %v", node.On)
-	}
-	if node.Using != nil {
-		buf.Myprintf(" using %v", node.Using)
-	}
 }
 
 func (node JoinCondition) walkSubtree(visit Visit) error {
@@ -2086,11 +1537,6 @@ const (
 	NaturalRightJoinStr = "natural right join"
 )
 
-// Format formats the node.
-func (node *JoinTableExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%v %s %v%v", node.LeftExpr, node.Join, node.RightExpr, node.Condition)
-}
-
 func (node *JoinTableExpr) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -2115,17 +1561,6 @@ const (
 	IgnoreStr = "ignore "
 	ForceStr  = "force "
 )
-
-// Format formats the node.
-func (node *IndexHints) Format(buf *TrackedBuffer) {
-	buf.Myprintf(" %sindex ", node.Type)
-	prefix := "("
-	for _, n := range node.Indexes {
-		buf.Myprintf("%s%v", prefix, n)
-		prefix = ", "
-	}
-	buf.Myprintf(")")
-}
 
 func (node *IndexHints) walkSubtree(visit Visit) error {
 	if node == nil {
@@ -2158,14 +1593,6 @@ func NewWhere(typ string, expr Expr) *Where {
 		return nil
 	}
 	return &Where{Type: typ, Expr: expr}
-}
-
-// Format formats the node.
-func (node *Where) Format(buf *TrackedBuffer) {
-	if node == nil || node.Expr == nil {
-		return
-	}
-	buf.Myprintf(" %s %v", node.Type, node.Expr)
 }
 
 func (node *Where) walkSubtree(visit Visit) error {
@@ -2453,15 +1880,6 @@ func replaceExprs(from, to Expr, exprs ...*Expr) bool {
 // It's not a valid expression because it's not parenthesized.
 type Exprs []Expr
 
-// Format formats the node.
-func (node Exprs) Format(buf *TrackedBuffer) {
-	var prefix string
-	for _, n := range node {
-		buf.Myprintf("%s%v", prefix, n)
-		prefix = ", "
-	}
-}
-
 func (node Exprs) walkSubtree(visit Visit) error {
 	for _, n := range node {
 		if err := Walk(visit, n); err != nil {
@@ -2469,11 +1887,6 @@ func (node Exprs) walkSubtree(visit Visit) error {
 		}
 	}
 	return nil
-}
-
-// Format formats the node.
-func (node *AndExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%v and %v", node.Left, node.Right)
 }
 
 func (node *AndExpr) walkSubtree(visit Visit) error {
@@ -2491,11 +1904,6 @@ func (node *AndExpr) replace(from, to Expr) bool {
 	return replaceExprs(from, to, &node.Left, &node.Right)
 }
 
-// Format formats the node.
-func (node *OrExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%v or %v", node.Left, node.Right)
-}
-
 func (node *OrExpr) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -2511,11 +1919,6 @@ func (node *OrExpr) replace(from, to Expr) bool {
 	return replaceExprs(from, to, &node.Left, &node.Right)
 }
 
-// Format formats the node.
-func (node *NotExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("not %v", node.Expr)
-}
-
 func (node *NotExpr) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -2528,11 +1931,6 @@ func (node *NotExpr) walkSubtree(visit Visit) error {
 
 func (node *NotExpr) replace(from, to Expr) bool {
 	return replaceExprs(from, to, &node.Expr)
-}
-
-// Format formats the node.
-func (node *ParenExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("(%v)", node.Expr)
 }
 
 func (node *ParenExpr) walkSubtree(visit Visit) error {
@@ -2567,14 +1965,6 @@ const (
 	JSONExtractOp        = "->"
 	JSONUnquoteExtractOp = "->>"
 )
-
-// Format formats the node.
-func (node *ComparisonExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%v %s %v", node.Left, node.Operator, node.Right)
-	if node.Escape != nil {
-		buf.Myprintf(" escape %v", node.Escape)
-	}
-}
 
 func (node *ComparisonExpr) walkSubtree(visit Visit) error {
 	if node == nil {
@@ -2624,11 +2014,6 @@ const (
 	NotBetweenStr = "not between"
 )
 
-// Format formats the node.
-func (node *RangeCond) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%v %s %v and %v", node.Left, node.Operator, node.From, node.To)
-}
-
 func (node *RangeCond) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -2655,11 +2040,6 @@ const (
 	IsNotFalseStr = "is not false"
 )
 
-// Format formats the node.
-func (node *IsExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%v %s", node.Expr, node.Operator)
-}
-
 func (node *IsExpr) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -2672,11 +2052,6 @@ func (node *IsExpr) walkSubtree(visit Visit) error {
 
 func (node *IsExpr) replace(from, to Expr) bool {
 	return replaceExprs(from, to, &node.Expr)
-}
-
-// Format formats the node.
-func (node *ExistsExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("exists %v", node.Subquery)
 }
 
 func (node *ExistsExpr) walkSubtree(visit Visit) error {
@@ -2764,24 +2139,6 @@ func NewValArg(in []byte) *SQLVal {
 	return &SQLVal{Type: ValArg, Val: in}
 }
 
-// Format formats the node.
-func (node *SQLVal) Format(buf *TrackedBuffer) {
-	switch node.Type {
-	case StrVal:
-		sqltypes.MakeTrusted(sqltypes.VarBinary, node.Val).EncodeSQL(buf)
-	case IntVal, FloatVal, HexNum:
-		buf.Myprintf("%s", []byte(node.Val))
-	case HexVal:
-		buf.Myprintf("X'%s'", []byte(node.Val))
-	case BitVal:
-		buf.Myprintf("B'%s'", []byte(node.Val))
-	case ValArg:
-		buf.WriteArg(string(node.Val))
-	default:
-		panic("unexpected")
-	}
-}
-
 func (node *SQLVal) walkSubtree(visit Visit) error {
 	return nil
 }
@@ -2800,11 +2157,6 @@ func (node *SQLVal) HexDecode() ([]byte, error) {
 	return dst, err
 }
 
-// Format formats the node.
-func (node *NullVal) Format(buf *TrackedBuffer) {
-	buf.Myprintf("null")
-}
-
 func (node *NullVal) walkSubtree(visit Visit) error {
 	return nil
 }
@@ -2813,29 +2165,12 @@ func (node *NullVal) replace(from, to Expr) bool {
 	return false
 }
 
-// Format formats the node.
-func (node BoolVal) Format(buf *TrackedBuffer) {
-	if node {
-		buf.Myprintf("true")
-	} else {
-		buf.Myprintf("false")
-	}
-}
-
 func (node BoolVal) walkSubtree(visit Visit) error {
 	return nil
 }
 
 func (node BoolVal) replace(from, to Expr) bool {
 	return false
-}
-
-// Format formats the node.
-func (node *ColName) Format(buf *TrackedBuffer) {
-	if !node.Qualifier.IsEmpty() {
-		buf.Myprintf("%v.", node.Qualifier)
-	}
-	buf.Myprintf("%v", node.Name)
 }
 
 func (node *ColName) walkSubtree(visit Visit) error {
@@ -2866,11 +2201,6 @@ func (ValTuple) iColTuple()  {}
 func (*Subquery) iColTuple() {}
 func (ListArg) iColTuple()   {}
 
-// Format formats the node.
-func (node ValTuple) Format(buf *TrackedBuffer) {
-	buf.Myprintf("(%v)", Exprs(node))
-}
-
 func (node ValTuple) walkSubtree(visit Visit) error {
 	return Walk(visit, Exprs(node))
 }
@@ -2882,11 +2212,6 @@ func (node ValTuple) replace(from, to Expr) bool {
 		}
 	}
 	return false
-}
-
-// Format formats the node.
-func (node *Subquery) Format(buf *TrackedBuffer) {
-	buf.Myprintf("(%v)", node.Select)
 }
 
 func (node *Subquery) walkSubtree(visit Visit) error {
@@ -2901,11 +2226,6 @@ func (node *Subquery) walkSubtree(visit Visit) error {
 
 func (node *Subquery) replace(from, to Expr) bool {
 	return false
-}
-
-// Format formats the node.
-func (node ListArg) Format(buf *TrackedBuffer) {
-	buf.WriteArg(string(node))
 }
 
 func (node ListArg) walkSubtree(visit Visit) error {
@@ -2930,11 +2250,6 @@ const (
 	ShiftLeftStr  = "<<"
 	ShiftRightStr = ">>"
 )
-
-// Format formats the node.
-func (node *BinaryExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%v %s %v", node.Left, node.Operator, node.Right)
-}
 
 func (node *BinaryExpr) walkSubtree(visit Visit) error {
 	if node == nil {
@@ -2962,15 +2277,6 @@ const (
 	Utf8mb4Str = "_utf8mb4 "
 )
 
-// Format formats the node.
-func (node *UnaryExpr) Format(buf *TrackedBuffer) {
-	if _, unary := node.Expr.(*UnaryExpr); unary {
-		buf.Myprintf("%s %v", node.Operator, node.Expr)
-		return
-	}
-	buf.Myprintf("%s%v", node.Operator, node.Expr)
-}
-
 func (node *UnaryExpr) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -2985,11 +2291,6 @@ func (node *UnaryExpr) replace(from, to Expr) bool {
 	return replaceExprs(from, to, &node.Expr)
 }
 
-// Format formats the node.
-func (node *IntervalExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("interval %v %s", node.Expr, node.Unit)
-}
-
 func (node *IntervalExpr) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -3002,11 +2303,6 @@ func (node *IntervalExpr) walkSubtree(visit Visit) error {
 
 func (node *IntervalExpr) replace(from, to Expr) bool {
 	return replaceExprs(from, to, &node.Expr)
-}
-
-// Format formats the node.
-func (node *TimestampFuncExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%s(%s, %v, %v)", node.Name, node.Unit, node.Expr1, node.Expr2)
 }
 
 func (node *TimestampFuncExpr) walkSubtree(visit Visit) error {
@@ -3030,11 +2326,6 @@ func (node *TimestampFuncExpr) replace(from, to Expr) bool {
 	return false
 }
 
-// Format formats the node.
-func (node *CurTimeFuncExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%s(%v)", node.Name.String(), node.Fsp)
-}
-
 func (node *CurTimeFuncExpr) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -3049,11 +2340,6 @@ func (node *CurTimeFuncExpr) replace(from, to Expr) bool {
 	return replaceExprs(from, to, &node.Fsp)
 }
 
-// Format formats the node.
-func (node *CollateExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%v collate %s", node.Expr, node.Charset)
-}
-
 func (node *CollateExpr) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -3066,21 +2352,6 @@ func (node *CollateExpr) walkSubtree(visit Visit) error {
 
 func (node *CollateExpr) replace(from, to Expr) bool {
 	return replaceExprs(from, to, &node.Expr)
-}
-
-// Format formats the node.
-func (node *FuncExpr) Format(buf *TrackedBuffer) {
-	var distinct string
-	if node.Distinct {
-		distinct = "distinct "
-	}
-	if !node.Qualifier.IsEmpty() {
-		buf.Myprintf("%v.", node.Qualifier)
-	}
-	// Function names should not be back-quoted even
-	// if they match a reserved word. So, print the
-	// name as is.
-	buf.Myprintf("%s(%s%v)", node.Name.String(), distinct, node.Exprs)
 }
 
 func (node *FuncExpr) walkSubtree(visit Visit) error {
@@ -3133,11 +2404,6 @@ func (node *FuncExpr) IsAggregate() bool {
 	return Aggregates[node.Name.Lowered()]
 }
 
-// Format formats the node
-func (node *GroupConcatExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("group_concat(%s%v%v%s)", node.Distinct, node.Exprs, node.OrderBy, node.Separator)
-}
-
 func (node *GroupConcatExpr) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -3167,11 +2433,6 @@ func (node *GroupConcatExpr) replace(from, to Expr) bool {
 	return false
 }
 
-// Format formats the node.
-func (node *ValuesFuncExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("values(%v)", node.Name)
-}
-
 func (node *ValuesFuncExpr) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -3184,22 +2445,6 @@ func (node *ValuesFuncExpr) walkSubtree(visit Visit) error {
 
 func (node *ValuesFuncExpr) replace(from, to Expr) bool {
 	return false
-}
-
-// Format formats the node.
-func (node *SubstrExpr) Format(buf *TrackedBuffer) {
-	var val interface{}
-	if node.Name != nil {
-		val = node.Name
-	} else {
-		val = node.StrVal
-	}
-
-	if node.To == nil {
-		buf.Myprintf("substr(%v, %v)", val, node.From)
-	} else {
-		buf.Myprintf("substr(%v, %v, %v)", val, node.From, node.To)
-	}
 }
 
 func (node *SubstrExpr) replace(from, to Expr) bool {
@@ -3218,11 +2463,6 @@ func (node *SubstrExpr) walkSubtree(visit Visit) error {
 	)
 }
 
-// Format formats the node.
-func (node *ConvertExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("convert(%v, %v)", node.Expr, node.Type)
-}
-
 func (node *ConvertExpr) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -3236,11 +2476,6 @@ func (node *ConvertExpr) walkSubtree(visit Visit) error {
 
 func (node *ConvertExpr) replace(from, to Expr) bool {
 	return replaceExprs(from, to, &node.Expr)
-}
-
-// Format formats the node.
-func (node *ConvertUsingExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("convert(%v using %s)", node.Expr, node.Type)
 }
 
 func (node *ConvertUsingExpr) walkSubtree(visit Visit) error {
@@ -3272,21 +2507,6 @@ const (
 	CharsetStr      = "charset"
 )
 
-// Format formats the node.
-func (node *ConvertType) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%s", node.Type)
-	if node.Length != nil {
-		buf.Myprintf("(%v", node.Length)
-		if node.Scale != nil {
-			buf.Myprintf(", %v", node.Scale)
-		}
-		buf.Myprintf(")")
-	}
-	if node.Charset != "" {
-		buf.Myprintf("%s %s", node.Operator, node.Charset)
-	}
-}
-
 func (node *ConvertType) walkSubtree(visit Visit) error {
 	return nil
 }
@@ -3298,11 +2518,6 @@ const (
 	NaturalLanguageModeWithQueryExpansionStr = " in natural language mode with query expansion"
 	QueryExpansionStr                        = " with query expansion"
 )
-
-// Format formats the node
-func (node *MatchExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("match(%v) against (%v%s)", node.Columns, node.Expr, node.Option)
-}
 
 func (node *MatchExpr) walkSubtree(visit Visit) error {
 	if node == nil {
@@ -3326,21 +2541,6 @@ func (node *MatchExpr) replace(from, to Expr) bool {
 		}
 	}
 	return replaceExprs(from, to, &node.Expr)
-}
-
-// Format formats the node.
-func (node *CaseExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("case ")
-	if node.Expr != nil {
-		buf.Myprintf("%v ", node.Expr)
-	}
-	for _, when := range node.Whens {
-		buf.Myprintf("%v ", when)
-	}
-	if node.Else != nil {
-		buf.Myprintf("else %v ", node.Else)
-	}
-	buf.Myprintf("end")
 }
 
 func (node *CaseExpr) walkSubtree(visit Visit) error {
@@ -3367,25 +2567,12 @@ func (node *CaseExpr) replace(from, to Expr) bool {
 	return replaceExprs(from, to, &node.Expr, &node.Else)
 }
 
-// Format formats the node.
-func (node *Default) Format(buf *TrackedBuffer) {
-	buf.Myprintf("default")
-	if node.ColName != "" {
-		buf.Myprintf("(%s)", node.ColName)
-	}
-}
-
 func (node *Default) walkSubtree(visit Visit) error {
 	return nil
 }
 
 func (node *Default) replace(from, to Expr) bool {
 	return false
-}
-
-// Format formats the node.
-func (node *When) Format(buf *TrackedBuffer) {
-	buf.Myprintf("when %v then %v", node.Cond, node.Val)
 }
 
 func (node *When) walkSubtree(visit Visit) error {
@@ -3402,15 +2589,6 @@ func (node *When) walkSubtree(visit Visit) error {
 // GroupBy represents a GROUP BY clause.
 type GroupBy []Expr
 
-// Format formats the node.
-func (node GroupBy) Format(buf *TrackedBuffer) {
-	prefix := " group by "
-	for _, n := range node {
-		buf.Myprintf("%s%v", prefix, n)
-		prefix = ", "
-	}
-}
-
 func (node GroupBy) walkSubtree(visit Visit) error {
 	for _, n := range node {
 		if err := Walk(visit, n); err != nil {
@@ -3422,15 +2600,6 @@ func (node GroupBy) walkSubtree(visit Visit) error {
 
 // OrderBy represents an ORDER By clause.
 type OrderBy []*Order
-
-// Format formats the node.
-func (node OrderBy) Format(buf *TrackedBuffer) {
-	prefix := " order by "
-	for _, n := range node {
-		buf.Myprintf("%s%v", prefix, n)
-		prefix = ", "
-	}
-}
 
 func (node OrderBy) walkSubtree(visit Visit) error {
 	for _, n := range node {
@@ -3453,22 +2622,6 @@ const (
 	DescScr = "desc"
 )
 
-// Format formats the node.
-func (node *Order) Format(buf *TrackedBuffer) {
-	if node, ok := node.Expr.(*NullVal); ok {
-		buf.Myprintf("%v", node)
-		return
-	}
-	if node, ok := node.Expr.(*FuncExpr); ok {
-		if node.Name.Lowered() == "rand" {
-			buf.Myprintf("%v", node)
-			return
-		}
-	}
-
-	buf.Myprintf("%v %s", node.Expr, node.Direction)
-}
-
 func (node *Order) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -3482,18 +2635,6 @@ func (node *Order) walkSubtree(visit Visit) error {
 // Limit represents a LIMIT clause.
 type Limit struct {
 	Offset, Rowcount Expr
-}
-
-// Format formats the node.
-func (node *Limit) Format(buf *TrackedBuffer) {
-	if node == nil {
-		return
-	}
-	buf.Myprintf(" limit ")
-	if node.Offset != nil {
-		buf.Myprintf("%v, ", node.Offset)
-	}
-	buf.Myprintf("%v", node.Rowcount)
 }
 
 func (node *Limit) walkSubtree(visit Visit) error {
@@ -3510,15 +2651,6 @@ func (node *Limit) walkSubtree(visit Visit) error {
 // Values represents a VALUES clause.
 type Values []ValTuple
 
-// Format formats the node.
-func (node Values) Format(buf *TrackedBuffer) {
-	prefix := "values "
-	for _, n := range node {
-		buf.Myprintf("%s%v", prefix, n)
-		prefix = ", "
-	}
-}
-
 func (node Values) walkSubtree(visit Visit) error {
 	for _, n := range node {
 		if err := Walk(visit, n); err != nil {
@@ -3530,15 +2662,6 @@ func (node Values) walkSubtree(visit Visit) error {
 
 // UpdateExprs represents a list of update expressions.
 type UpdateExprs []*UpdateExpr
-
-// Format formats the node.
-func (node UpdateExprs) Format(buf *TrackedBuffer) {
-	var prefix string
-	for _, n := range node {
-		buf.Myprintf("%s%v", prefix, n)
-		prefix = ", "
-	}
-}
 
 func (node UpdateExprs) walkSubtree(visit Visit) error {
 	for _, n := range node {
@@ -3555,11 +2678,6 @@ type UpdateExpr struct {
 	Expr Expr
 }
 
-// Format formats the node.
-func (node *UpdateExpr) Format(buf *TrackedBuffer) {
-	buf.Myprintf("%v = %v", node.Name, node.Expr)
-}
-
 func (node *UpdateExpr) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -3573,15 +2691,6 @@ func (node *UpdateExpr) walkSubtree(visit Visit) error {
 
 // SetExprs represents a list of set expressions.
 type SetExprs []*SetExpr
-
-// Format formats the node.
-func (node SetExprs) Format(buf *TrackedBuffer) {
-	var prefix string
-	for _, n := range node {
-		buf.Myprintf("%s%v", prefix, n)
-		prefix = ", "
-	}
-}
 
 func (node SetExprs) walkSubtree(visit Visit) error {
 	for _, n := range node {
@@ -3612,19 +2721,6 @@ const (
 	TxReadWrite = "read write"
 )
 
-// Format formats the node.
-func (node *SetExpr) Format(buf *TrackedBuffer) {
-	// We don't have to backtick set variable names.
-	if node.Name.EqualString("charset") || node.Name.EqualString("names") {
-		buf.Myprintf("%s %v", node.Name.String(), node.Expr)
-	} else if node.Name.EqualString(TransactionStr) {
-		sqlVal := node.Expr.(*SQLVal)
-		buf.Myprintf("%s %s", node.Name.String(), strings.ToLower(string(sqlVal.Val)))
-	} else {
-		buf.Myprintf("%s = %v", node.Name.String(), node.Expr)
-	}
-}
-
 func (node *SetExpr) walkSubtree(visit Visit) error {
 	if node == nil {
 		return nil
@@ -3638,14 +2734,6 @@ func (node *SetExpr) walkSubtree(visit Visit) error {
 
 // OnDup represents an ON DUPLICATE KEY clause.
 type OnDup UpdateExprs
-
-// Format formats the node.
-func (node OnDup) Format(buf *TrackedBuffer) {
-	if node == nil {
-		return
-	}
-	buf.Myprintf(" on duplicate key update %v", UpdateExprs(node))
-}
 
 func (node OnDup) walkSubtree(visit Visit) error {
 	return Walk(visit, UpdateExprs(node))
@@ -3666,11 +2754,6 @@ func NewColIdent(str string) ColIdent {
 	return ColIdent{
 		val: str,
 	}
-}
-
-// Format formats the node.
-func (node ColIdent) Format(buf *TrackedBuffer) {
-	formatID(buf, node.val, node.Lowered())
 }
 
 func (node ColIdent) walkSubtree(visit Visit) error {
@@ -3744,11 +2827,6 @@ type TableIdent struct {
 // NewTableIdent creates a new TableIdent.
 func NewTableIdent(str string) TableIdent {
 	return TableIdent{v: str}
-}
-
-// Format formats the node.
-func (node TableIdent) Format(buf *TrackedBuffer) {
-	formatID(buf, node.v, strings.ToLower(node.v))
 }
 
 func (node TableIdent) walkSubtree(visit Visit) error {
